Store fitness for each member of population
This commit changes the data structure used to represent a population
from a simple set of members to a map from members to fitnesses. This
prevents (some) needlessly repeated calculations of fitnesses of DFAs.

Performance should be significantly improved, although `engineer` still
doesn't seem to be able to produce a 1.0-accuracy DFA for
bit-counting-predicate.